Support async read from clipboard
In case that read doesn't returns the data immediately, it needs to move
the data from port/state/data to port/data for later reading.

Currently the implementation assumes that the clipboard can be read in
one read when it's ready, which means a partial clipboard read is not
allowed.