[Linux] Optimize window moving
Because X sends out MOVE and RESIZE in the same ConfigurationNotify
event, the application needs to remember the old window size to avoid
doing unneeded resizing. The old size in gob can't be used because the
ConfigurationNotify could be actually caused by size changes in Rebol
script and then by OS_Update_Window, which updates the old size to the
new size, so when the ConfigurationNotify event arrives the old size was
already updated.

This also fixes an infinite loop caused by:
1. wmctrl change the window to fullscreen
2. a RESIZE with the screen size sent to rebol script
3. rebol script sets a wrong window size, because it assumes "window
size" is smaller than the "work area" size
4. OS_Update_Window is called with the new window size
5. Window manager refuses to re-size the window, because the fullscreen
flag is set on this window
6. a new RESIZE event is sent, because the window size returned from X
server is different from the one saved in gob

This fixes it by not using the old size in gob, but in a dedicated old
size purely for re-sizing detection