Fix a crash on 64-bit systems
Word hashes are supposed to be an array of 32-bit integers (See
Make_Word in c-word.c). A mismatch between Init_Words and Expand_Hash
makes it crash when it clearing new hashes (trying to clear more space
than it allocates). To make it obvious, change WORD_TABLE_SIZE to 1 at
the beginning of c-word.c