[Linux] Try to compress RESIZE events
If the user resizes the window very fast, a series of Resize events
would be received, there is not need to resize and recompose the window
for every such event. Save them in an array and resize them to the final
size only.

An expose event will follow each resize event, so it needs to be detected
and ignored.

How many resize events can be compressed, would depend on the rendering
and resizing speed.