Copy the whole uinon in the series
On 64-bit systems, "size" is 32-bit long, and the anonymous union is
64-bit long, just copying "size" in Prop_Series is not enough. Adding a
new field "all" in the union for this purpose.

This fixes a crash appening a value to a map!. Because the map! uses
"series" in the union, and Prop_Series only copies half of it (by
copying the "size" member), later access to "series" results in a
segment fault