Allow direct memory access from struct!
An attribute block can be added in the beginning of a struct! block,
which includes these settable attributes:
 * raw-memory: the memory location of the struct! storage
 * raw-size: the size of the "raw-memory", this will be checked against
the definition of the struct.

Examples:
a: make struct! [
	[raw-memory: 10000]
	uint32 i
]

The whole struct will be stored in the momery at 10000. Reading "a/i"
will retrieve the content in the address 10000, and changing "a/i"
will change its content.

This is useful in the callback from C (Not implemented yet) in this
scenario:
In C,
<code>
   void func (void (*cb) (void *, size_t),
	      void *data,
	      size_t size)
{
	cb (data, size);
}
</code>

In Rebol,
<code>
func: make routine! [
	[
		cb [pointer]
		data [pointer]
		size [int64] ; or int32 on 32-bit systems
	]
	lib "func"
]

cb: mk-cb [
	[
		data [pointer]
		size [int64] ; or int32 on 32-bit systems
	]
][
	s: make struct! compose/deep [
		[
			raw-memory: (data)
                        raw-size: (size)
		]
		int32 i
		int32 j
	]

	;s/i will refer to "a/i" below
	print ["i:" s/i] ;output 100
	print ["j:" s/j] ;output 200
]

a: make struct! [int32 i: 100 int32 j: 200]
func (reflect cb 'addr) (reflect a 'addr) (length? a)
</code>