Use heuristic to adjust the frequency of recycle
When ballast is too small, a lot of CPU will be wasted on recycling.
When it's too big, memory will be wasted. As the system keeps tracking
the memory used/recycled, use it as the indicator to determine how much
memory the application needs.

The idea is: if the GC_Ballast after a recycle is lower than half of the
ballast, which indicates not a lot of memory can be recycled, increase
the ballast by half, such that it approximately has a whole size of
"ballast" of memory to use before next recycle happens; if GC_Ballast after
a recyle is as twice big as the ballast, which indicates more memory can
be recycled, reduce the ballast by half, so recycle runs more often.

This largely reduces the startup time of app that requires huge memory