Fix a crash when pasting from clipboard
Because clipboard operation on Linux is asynchronous, a READ event is
issued, which results in a "UPDATE" call on the port, so the req->data
can't be free'd when A_READ is returned. Move this to handler of
A_UPDATE. As a CLOSE event could follow READ, and it will also cause an
"UPDATE" call, so check for req->data being NULL to avoid confusion.