Do not use the system qsort
The implementation varies on different platforms (it seems to be stable
sort on Linux, while unstable on Windows).

sort ["a" "A"] gives different result on these platforms when using the
system qsort

Rename it to reb_qsort to avoid conflict when linking. It seems to be
easy to redefine the system function, but not as easy for visual studio.
As its usage is quite limited, it's easier to rename it.