Take binary! for I/O redirection in CALL
The difference between string! and binary! is that, with string!, the
interpreter will try to do the encoding conversion before/after passing
in/out of the command; with binary!, no such conversion is done in the
interpter, everything will be passed in/out as-is. So if the conversion
is not desirable/correct, use binary! (and do the conversion yourself).