Add a new type rebval to struct!
This type is quite useful to pass rebol values to a callback functions.
It's quite common in C to have this kind of functions:

	void foo(void (*bar)(void *data), void *user_data);

Inside foo, it does:

	bar(user_data);

Their rebol wrappers are:

	foo: make routine! compose [[
		bar 		[pointer]
		user-data 	[pointer]
		return: 	[void]
	] (lib) "foo"]

	bar: func [
		data [integer!]
	][
	]

	bar-cb: make callback! reduce [[
		data [pointer]
		return: [void]
	] :bar]

So far, everything looks fine. What's not so fine is when foo is called
and some user-data needs to be passed to bar. Because user-data is a
pointer, it can only be an integer, or result from "addr-of struct!".
Passing in an integer is trivial, and the usage in "bar" is simple.
Most situations requires passing in complex data structure, which is the
result form "add-of". However, as struct! is designed to solely work with
C datatypes, it only supports basic atomic types or another "struct!",
which is not quite convienent to work with, especially when it involves
dynamic memory management.

Say you want "bar" to put some data into an array, and the length of
data is unknown before hand. Manual calls to "malloc" or "free" might be
required, which makes "bar" not really like a REBOL function anymore:

	foobar: make struct! [
		uint32	len ;
		pointer addr ;since the length is unknown, it gotta be a
pointer
	]
	foo addr-of foobar

and:
	bar: func [
		data [integer!]
		/local foobar
	][
		foobar: make struct! compose/deep [
			[raw-memory: (data)]
			uint32	len
			pointer addr
		]
		;foobar/addr: malloc ...
		;free foobar/addr
	]

This is where "REBVAL" comes to rescue.

The usage is quite simple, just use it inside struct! as any other
atomic datatype:
	foobar: make struct! [
		rebval	v
	]
	;foobar/v can be any rebol value, not just those with atomic C types

	foobar/v: copy []

	;it can also be any of these:
	;foobar/v: "hello"
	;foobar/v: make object! [i: 0]

and it can be passed to foo:

	foo addr-of foobar

inside bar:
	bar: func [
		data [integer!]
		/local foobar
	][
	    	foobar: make struct! compose/deep [
			[raw-memory: (data)]
		    	rebval	v
	    	]
		;now foobar/v will be whatever is set before "foo" is
being called.
		;and you can handle it as any other rebol values:
		; object? foobar/v

		append foobar/v "hello"
	]

and after the "foo" call, foobar/v becomes ["hello"], no manual memory
management at all.