Release extra memory as early as possible
Memory allocated while calling the routine should be free'ed right after the
calling is done. QUEUE_EXTRA_MEM saves the memory to the routine struct,
and it will not be free'd until the routine is free'd. If the routine is
called repeatedly, the memory could increase each time. This is not a
strict memory leak (as the memory is still reachable, and will be
eventually free'd), but has an equivalent effect as a memory leak.