assert()+Panic()+Trap() revamp for debug build
This is a comprehensive change-over for Rebol to use C's <assert.h>
and to separate assertions out from "Panics" and "Traps".  The
former definition of ASSERT in Rebol would include those asserts
in release builds (depending on the "assert level") and required
the inclusion of a "Rebol panic code".

The comments added to sys-core.h summarize the situation, and the
most observable effect (besides renaming Crash to Panic) is the
versions of Trap and Panic that indicate a DEAD_END condition.
This informs the compiler that the lack of a return on that
particular code path is intentional.  This becomes particularly
important in stronger path checking than GCC or Clang provide
by default.