Fix variable wchar_t size issues with wcslen()
wchar_t has no guaranteed size on non-Windows platforms, and in fact
I discovered on my 64-bit Linux that sizeof(wchar_t)==4.  This means
there needed to be a UCS-2/UTF-16 length checker that could handle
REBUNI.  This adds Strlen_Uni for that functionality and eliminates
the calls to wcslen().