char*/REBYTE fixes with casts-for-the-masses
This commit is the logical consequences of turning on -Wpointer-size
so that REBYTE* and char* become incompatible types, and turning all
the C literal constants into 'const char *'.  Whenever a point is
reached that requires a transition, it tends to be meaningful and
worth paying attention to--as a switch from unencoded characters to
encoded UTF-8 bytes (or similar).  I wrote an article about it, which
provides a thorough overview:

    http://blog.hostilefork.com/c-casts-for-the-masses/

Not only does it improve the robustness of the code and catch potential
misunderstandings and bugs...but I think it looks better too.