Hack to get around argv[] -Wcast-qual warning in C
Although in "casts for the masses" I describe a way to leverage a C++
build to make sure your C code isn't casting away consts, it turns out
that C has a compiler warning: -Wcast-qual which can rigorously demand
you never cast away a const.  While that is going to not work for many
such codebases, Rebol can actually do it... but there's one problem
here with passing argv to execvp.  It expects char * const *, and the
desire is to pass in 'const char**'.

If it weren't just this *one* case of a warning, and they were all
over the place and unfixable, I'd say give up on -Wcast-qual and just
let the C++ build check it.  But with a single case serving as an
opportunity to turn on -Wcast-qual and -Werror to keep new instances
of casting-away const from popping up from C contributors...I feel
it's worth it to tunnel under the cast and keep the warning on.