Quarantine REBCHR as opaque in core
I describe in make-os-ext.r the gripe with REBCHR and how it seems
like a bad idea in most all the places it occurs.  A better story is
needed, but while we wait for that better story it's good to have it
quarantined and type-checked as well as possible.  This limits the
reach of REBCHR by making it a struct that can't be easily used as
a character when you're inside the core.  And in the host, it is
defined as its native type...so the host code doesn't talk about
REBCHR anywhere but the function prototypes.  It can then just pass
those to things taking the native type with no cast.

So from here on out: if one is writing Windows host code, just use
wchar_t.  And if it's other clients, just use char.  It makes more
sense and will fit in with whatever the longer-term solution is.

Also, this eliminates TCHAR usage.  TCHAR is Windows' own version of
REBCHR; a conditionally compiled type which could be a char or a
whchar_t.  It is considered obsolete, and versions of Windows that
don't have unicode support are hard to come by.  Even Windows 9x
had support retrofitted.

Point being that using TCHAR in an isolated part of the hostkit
isn't going to make Rebol work on Windows 3.1, given all the other
uses of wchar_t.  If one wanted to compile a non-GUI executable for
older Windows or DOS, one should do it with DJGPP with POSIX support.