Patches for turning on -Wparentheses
Rebol had a macro NZ(...) for wrapping up assignments to indicate you
really meant to do an assignment and not test for equality.  But as
idioms go, I think the double-parentheses is a bit less jarring while
still calling attention to the "I didn't mean this to be an equality
test" idea:

    http://stackoverflow.com/a/5476795/211160

The NZ not-equal-zero macro is weird, and if there's going to be a
removal from programming in C and saying '!= 0' then that invites the
creation of macros for UNLESS and EITHER or all kinds of redesign of
C.  So I killed NZ and moved everything to the double-parentheses
convention, given that it is so common that there's even a warning
where the compiler checks it for you.