Patches for turning on -Wcomment and tidy a-lib.c
This fixes a few malformed comments that wanted to put /* in the
comment text itself.  It also tidies the headers in a-lib.c

Note comment headers with the asterisks in a certain form are what are
used to cause Rebol to gather up functions in tmp-funcs.h.  (They are
the reason that 'make prep' can keep you from having to add functions
to a header file each time you put something in one C file you want
to call from another.)

However, the rule for matching to get that export is simply:

    thru "/******" to newline
    "^/**" to newline
    "^/*/" any [#" " | #"^-"]

So you can get away with some pretty ugly comments.  Also, sometimes
you don't need to export things because they're found by another
process.  In the case of the API, it's found by the RL_API tag in
make-reb-lib.r.  So unless Rebol were to want to call the RL_API
directly, it wouldn't need the headers.

Less ugly than malforming the export comments is to let make-headers.r
be cued by the RL_API that you don't need an export (as it knows you
don't need one for static, for instance).  Then fix up the headers all
nicely.  So I did.