Don't use function debug byte casts outside Core
The function forms of s_cast, b_cast, cs_cast, and cb_cast are done as
typical exports from make-headers.r in Debug builds, to guarantee you
are passing the right type in.  (non-const REBYTE* goes into s_cast and
non-const char* comes out... const char* goes into cs_cast and const
REBYTE* comes out)

However, the tmp-funcs.h is not supposed to be included outside of
Core.  The alternative would be to make RL_s_cast, RL_b_cast, etc...
but then Rebol isn't supposed to be calling those!  It gets messy.

I could create C++11 versions that would work anywhere and not need
to link to a function to check the types.  But for the immediate moment
we save the debug check for the Core to avoid linking host to functions
it shouldn't be seeing.