Skip_To_Char => Skip_To_Byte (not inlined)
I needed to change and move this around anyway, because it was having
some problems with definitions as it was an attempt to inline a little
piece of code.  Inlining is tricky and the compiler is usually better
at it than you (the standard says it can ignore inlining if it wants
anyway, so you're only hinting).

If performance tests truly point to this as a smoking gun that is
not being optimized properly, we can revisit it.  But in trying to get
the C++ build linking it's easier to make it a function like any other.