Mark memset of a value "zero" more intentionally
I came across the tests for zero and greater than or less than zero,
and was a bit unnerved that it assumes that setting the bit pattern
of a value to all zero bits and then setting to the type is that type's
"zero value" to be compared against.  If that's the rule then it has
interesting implications for the value types.  But rather than just
leave it undocumented, I added a VAL_SET_ZEROED macro to call your
attention to it in the future.