Patches for turning on -Wsometimes-initialized
Think I got this one right in the vector case, don't know how to test it
so that will have to wait.  The other warning about initialization just
meant it needed to be told that a Trap won't return, so the compiler
doesn't worry that it can and have an uninitialized value used.