Distinguish codec actions from codec types
The prefixes of all-caps macros help you tell what they are for, but
the CODI_ prefix was used for both media types and codec actions.  It
seemed best to turn the codi actions into CODI_ACT_ prefix because
that is more descriptive, and leave the CODI_ for the media types.

(Assuming you have much faith in the existing codec model, which I
do not actually.  But it's there, it's running, might as well give it
some consideration in naming.)