Fix trailing commas in enumerated types
These aren't legal in C89, although interestingly enough they are legal
in C99.

The easiest way to fix many of these commas when they make sense is to
just add a _MAX as the last entry.  It doesn't need a comment, and a
generative script can just make it the last thing it does.  Bitmasks
or other things that wind up in enums don't make sense for that
approach, so you actually have to delete the comma.