Remove compound type and designated initializers
I hadn't even seen designated initializers before.  Kind of cool but
doesn't really add much to this case...arguably makes it a bit more
confusion.  Good to know about, but as they are -Wc99-extensions it
seems not worth it to break the C89 buildability for it.

(I'll point out though, that as host and core separate even more, it
will not be necessary to follow Core rules in hosts.  I of course want
C++ hosts...)