Replace STOID with static void
While this is not completely necessary, I have worked pretty hard on
this, and deserve a bit of satisfaction.  Killing STOID gives me some.

The preprocessor is used heavily by Rebol in places where it shouldn't
be, just because the preprocessor is what C has.  And I don't have a
problem with that...but I do think this kind of thing is not very
fitting.  If one wants to generate C code from a Rebol file, that's
okay...but trying to abbreviate C like this messes with readability
and breaks the structure of the code.