Fix passing 'long long*' to 'long*' routines
Where possible, the Atronix/Saphirion build uses native 64-bit math
functions.  However, on 64-bit systems there is a type conflict when
you pass a 'long long*' to a routine expecting a 'long*', even if
'long long' and 'long' are the same size.  They're still different
types, and since the math routine takes a 'long*' and REBI64 resolves
to a 'long long*', you need a cast on those.