Patches for turning on -Wchar-subscripts
I suppose it's a mistake that people will often use character subscripts
and not realize they're signed...and the domain of a character is small
enough that you may well be trying to build a table that uses all 255
values.  So telling you that you're using a signed character as an
array subscript seems a reasonable warning.  Not many cases in Rebol.