Standardize Put_Str interface to mutable REBYTE*
Put_Str isn't used very much, it's not very clear what it's for, and
it seems to be confused about whether it needs to allocate and free a
string or not.  I simplified this just to make it easier to get rid
of, since it's really unclear who should be using this or why.