Cast to get around enum-out-of-range check at boot
Because there is an enum type for REB_TYPE, and I had to reconcile
the type of its assignments, I went ahead and made the VAL_TYPE macro
cast the result to the enum type.  Which makes it smarter than a
number...if you test against a number not in the enum range it can
tell you.  The boot sequence deliberately puts a bad value into the
bits and then checks to make sure it can read them out.

Rather than having it write a known valid type into the bits, I just
left it as-is and cast around it.  There's no particularly good or
bad reason for having done so, but I had to do one or the other.