Use C linkage (no name-mangling) for externs
I find surprising behavior when trying to get the linker to work a lot
of the time, and so I don't really know if I've done this the best way.
It seems sometimes specifying extern "C" linkage is needed when I
would think if both the definition and the prototype were using the
C++ name mangling they'd match up just as well.  In any case, this gets
it to link, so we'll start here.