Get rid of anonymous unions (illegal in C99)
Anonymous unions are not legal in C99 forward, and they are illegal in
all C++ versions.  It can be a bit hard to come up with names, and I am
not saying I came up with the best ones.  Honestly I find the DEVREQ a
bit a strange idea; it doesn't feel all that much like a robust way of
doing extensions to just keep piling things into a centralized struct.

In any case, the names are at least distinct (requestee, eventee) and
can stand out to be changed later.  This is more just to get the ball
rolling so the C99 and C++ builds work.