ROUTINE!/STRUCT! pointer casting (including 32bit)
Broke this out separately from the rest of the code as it is
something @ShixinZeng is working on and wanted it to be isolated.

There's a warning when you are casting something that isn't the
size of a pointer to a pointer, and there's some code in ROUTINE!
that on 32-bit builds was using the low order bits of a 64-bit
value to hold pointers in that case...hence truncating.  This was
apparently the intention, so the way to get rid of the warning
is to cast to a REBUPT (unsigned integer for holding a pointer)
before casting to a pointer.

Some other basic casts are required for C++.