Remove #pragmas for disabling warnings
One really should not use pragmas if avoidable, as they are
implementation-defined.  (At one point in time gcc would shell out
and run a game of Tetris or whatever it could find if you used one,
to make a point).

If pragmas are to be used, they should be selectively applied to work
around an issue with a specific piece of code.  These blanket pragmas
that are included by every source file and apply over the whole
codebase are better (and more clearly) done at the makefile level.