Sync OSX and POSIX duplicated code
There's a very unfortunate duplication of code right now across hosts.
In the open source release there were two hostkits: POSIX and Windows.
Now there's:

 - Windows
 - Linux
 - OSX
 - POSIX
 - and "Porting Templates"

Much of the code is identical between Linux, OSX, and POSIX.  But the
files are just duplicated.  That porting templates is basically a copy
of the POSIX code again is even more frustrating.

Atronix has been maintaining Linux/Windows, so those are the ones that
work.  As a stop-gap measure (short of a redesign), this copies the
Linux changes as best possible to try and revive a POSIX and OSX build.

Now that we've seen what NOT to do, we can contemplate ideas for doing
this multi-platform model in a more sane way.