Revamp make-make-r and related scripts, less in reb-config.h
The 'make make' and 'make prep' processes weren't very clear, so this
updates them to make it easier to see what's going on.  Along with
comments and some better common routines, it now is able to clearly
use the auto-detect feature for make-make so it will work most of
the time without specifying an OS_ID (by just using the ID out of
the r3-make executable that is running).  While putting more switches
in the systems table made rows not fit on one 80-character line, I think
the resulting clarity is worth it, and we can have considerably less of
the configuration buried in reb-config.h