Adapt README.md from Coherence 1.5, make makefile
Wanted to put something better than "Coming Soon" for the README.md
so I borrowed from the Coherence 1.5 PR a bit, and wrote some more
stuff.  Also added some more helpful feedback to the makefiles.  I
think 'make make' is unnecessarily cryptic compared to 'make makefile'
and there's probably some better naming to be done than
'makefile.boot' (perhaps two makefiles… generate.makefile and
rebol.makefile depending on whether you want it to build or not?)  All
of this is still new so it's a matter of getting things started.