'make makefile' re-enabled for makefile.boot after SO answer
http://stackoverflow.com/questions/31490689/

Found out what was going wrong; makefile is a "specially treated" target
(I suspected this).  The special treatment is okay for makefile.boot but
not for the generated makefile; but there's no particular reason why you
should need two ways to generate makefiles.  Still, we'll keep the old
'make make' working for a bit as this shakes out.