Remove "length = 0 means calculate length" for Make_Word
Rebol has a number of places where it will calculate the length of things for
you if you pass in -1 or 0, otherwise it uses the length you pass in. This
kind of convention asks for errors, and runs up against questions of how to
catch the "empty word" and what to do about it.  Best to avoid it.