Switch version number to 2.102.0, for the time being
This back-pedalling (on the 3.0.99 version number currently used by
Atronix) intends to allow us to eventually do a "proper" 3.0.0 release,
with fanfare and stuff. It may not be the best of choices, but should be
sufficiently conservative to also allow us to change our mind later.
(NB: rebol/rebol currently uses 2.101.0.)