No POLLOUT on stdin of process for empty string
Hopeful fix to #21.  When you pass an empty string to the process
host for POSIX it was doing a poll() with POLLOUT, even though
there's no bytes to offer on its standard input port.  I gather this
is not a legal thing to do and you shouldn't set the flag if that is
the case.