Get release build working again (-DNDEBUG)
Missing semicolon, but more irritating is that you can't have an identifier
that is a macro in the release build but a function in the debug build.
You have to go through a layer of indirection, because the function
prototype scanner in `make prep` doesn't pay attention to #ifdefs.  This
means you'll have an unused function prototype that collides with the
macro.  A problem that might need to be looked at more generally, but
the easy thing here is just to move the function names in the debug
build out of the way (is an underscore at the end too subtle?  Should
it be s_cast_DEBUG?)