Series_Panic() trick for tracking problematic series
When you access an invalid memory location in Address Sanitizer and
Valgrind, they remember the stack where that memory was freed so it
can show you the smoking gun.  This was my idea for piggy-backing on
that trick to let you get the smoking gun of "where was this series
allocated and what was its stack then?"  In debug builds, a guard value
is allocated as an integer for each new series and then freed right
away.  If something about a series bugs you, then you call Panic_Series
and it will deliberately dereference the guard.  Now you see the stack
where the series was allocated and the stack where you are when the
questionable thing was noticed.  VERY helpful.