Add notice if R3_ALWAYS_MALLOC is on
I messed up typing `set R3_ALWAYS_MALLOC=1` and forgot what
OS I was on, where I have to type `export R3_ALWAYS_MALLOC=1`.
Added a debug print to figure out if it was on or not, and then realized
it's probably the kind of thing you would like to be reminded is on.  So
I decided to turn the temporary Debug_Str into a permanent banner.