Preliminary git-hooks/pre-commit to start the ball rolling
I found this on the web, and it doesn't check to enforce tab usage specifically,
it just notices mixtures in the diff of tabs and spaces.  That's a good enough
start, although it does mean you will possibly wind up getting warnings about
problems in the spacing that you didn't introduce but were there in the lines
near what you were committing.