PROBE with source location and va_list by pointer
I had a stupid mistake which involved calling a varargs function with the
wrong order of parameters, which led me down a path of discovering that
you aren't really (?) supposed to pass va_list by value.  Changing it didn't
fix my problem, but so long as I was there I left it in.  See:

http://stackoverflow.com/a/3369762/211160

This also adds a little static counter into Do_Next to catch an evaluation by
count, which can be helpful in debugging something deterministic, which
a lot of cases (like booting, and the tests so far) are.