Interim workaround for CC#2221
This does not solve the problem of how to get a good
metaprogramming interface for working with the bodies of
functions and closures, in terms of knowing what local is bound
where vs. what was unbound.  Getting to that point will require
some more work with bind? ('binding-of') and specifically a way
of interacting with closure 'instances' that can be distinct from a
closure 'archetype'.  But it should help for now.  Details in ticket.