assert in GC case of calling Panic_Core
The parameterized Panic routines don't assert with a file and line
number the way that Panic() does when it's a macro.  I'd put an
assert on all the cases by hand but missed this one.  Nice to have
a line number and file if it happens (and break into the debugger
in Debug builds vs. exiting peacefully)