Workaround for MOLD in early boot phases
This setting for checking the number of decimal digits you want to
have in MOLD output comes from configuration blocks in the form
of Rebol data, and isn't available when the program starts.  You
have to be up to the point where those blocks are loaded, or MOLD
will crash.  I don't know about such settings influencing the process
globally in general...seems to introduce unnecessary variability to
the process.  Should be a parameter to the specific mold going on,
at least.  But in order for debug output to display data, a conditional
check to keep it from crashing is necessary.