Change 'word' in Do_Next/Push_Func to 'label'
The 'word' variable is a bit generic in Do_Next as written and it's
not entirely clear what it does or how it works.  There's also some
obfuscation in terms of turning it into a symbol and various ways of
saying it's NULL.  This centralizes the NULL handling into Push_Func
and establishes a definition of what it's for, namely a 'label' for the
call that's shown when debugging, that is set if you do an invocation
through a word or path.