Respect protection w/GET_VAR+GET_MUTABLE_VAR
This narrows the various flavors of Get_Var down to two basic core
routines and the five macros you should actually use.  What this does
is makes it so that if you do not explicitly request mutable access on a
lookup of a word's value, you get a const pointer back.  This indicates
that you *did not check* for PROTECT status and it means the compiler
will complain if you try and pass that to a modifying operation.  If you
actually need a mutable reference, then you use GET_VAR_MUTABLE
and it will check and trap on an error if the variable is protected.

(There are also TRY_GET variants which will return NULL instead of
raising an error.)

The Get routines include more comments for following what is going
on, as well.