Remove /NOW refinement from QUIT (CC#1743)
There is a technical reason why I want to do this, but even before the
technical reason I didn't care for it.  As BrianH says this is an "arms
race" and QUIT should mean quit the interpreter now unless something
catches it explicitly.  The upshot of this is that if your script calls another
script and that script quits--and you do not explicitly CATCH that,
then you quit too.

Carl said it needed serious discussion, but said to refer to his
blog, which is down ATM.  It seems the general consensus has
been that /NOW should go, and Red's QUIT doesn't have it
either.