Discern conditional true/false from LOGIC true/false
There's a miniscule performance reason not to test something you already know
to be a LOGIC! to see if it is a LOGIC again, and test against none.  But a better
reason is just the clarity of the code so you understand which tests are for the
idea of "conditional truth" or "conditional falsehood" and which are actually
testing for TRUE-valued or FALSE-valued LOGIC!s.