Add new pseudo-type "TRASH!"
A trash value could really reuse a bit pattern out of pretty much any other
type, but for now it's just REB_MAX + 1.  The idea is that sometimes you
need to write a "valid" value into a space for a temporary amount of time
with the intent that it will be filled in later...and you could use UNSET! for
this purpose (function returns did) but then you wouldn't be able to tell
an intentional UNSET! from a function that accidentally didn't set the
return value at all.  With a special trash value that isn't actually used, you
can distinguish those cases.  There is "unsafe trash" which the GC will
not tolerate, and then "safe trash" which it will ignore.  A filename and
line number can be found in the debugger of where a trash value was
initialized, which can be helpful as well.