Distinguish + Tighten Append_Value/Append_Val
Among the "very confusing names" in Rebol are the two functions Append_Val
and Append_Value.  The former does what you might expect; if you have a
pointer to a REBVAL cell and a block series, it will append that value to that
series.  The latter--however--allocated a NONE! value at the tail of the series
and returned a pointer to it.

On an efficiency note, you often don't want a none.  In fact, most of the time
you don't...so Append_Value just had what it wrote overwritten.  So it makes
more sense to speak of it as an allocation.  Hence the new name and
function: Alloc_Tail_Blk.  With that rename done, Append_Val could be
renamed to Append_Value and things make more sense.

Because the new Alloc_Tail_Blk does not initialize (and actually sets the
bits to "Trash") I had to go through cases that might have depended on
the value being NONE!.  I didn't do an analysis up the call stack to
see if it *actually* needed them to be none, just if there wasn't an
initialization immediately following I erred on the side of caution and
set it to none to mimic the previous behavior.