Switch from GNU-style to POSIX-style declaration of environ
Quoting the Linux environ(5) manpage:

    Synopsis

    extern char **environ;

    Description

    The variable environ points to an array of strings called the
    'environment'. (This variable must be declared in the user program,
    but is declared in the header file unistd.h in case the header files
    came from libc4 or libc5, and in case they came from glibc and
    _GNU_SOURCE was defined.)

    (Source: http://linux.die.net/man/5/environ)

As we (unneccessarily) defined _GNU_SOURCE in
src/os/posix/host-config.c, the (also unnecessary) #include of unistd.h
brought in a declaration of `environ` for glibc-based systems (such as
many a Linux) which was missing on non-glibc systems (such as OS/X).

As nothing in src/os/posix/host-config.c needs _GNU_SOURCE defined, we
simply remove this define and switch back to the POSIX-standard way.