Batch of const-correctness changes (mostly MOLD-related)
This is another "const-correction" wave kept separate from less
mechanical changes in order to make review easier.

There are a couple of non-trivial points in this one, however.  One is
that there is an actual sneaky thing being done in the REBEVT
structure with a mutable string field that host code apparently does
an OS_ALLOC on a byte string of, later expecting the core to
do a conversion of that into a string-series REBSER.  It's effectively
a "mutable" field even in const scenarios.  Didn't look deeper
but there is hopefully a way to avoid this technique.

Then an issue with where THROW-style errors (that have no error
"object") would fake an object by mutating the error to have a
throw number and an object, and I have some changes that
don't let you do that.  Also Emit_Date seems to need to modify the
passed-in date for some reason, so we make a copy there.