Axe TOS1 protocol from Apply_Func*/Do_Sys_Func
In order to "simplify" the protocol for certain clients, there was a pattern
in the code where after the stack had been popped some routines would
return a pointer to the previous (now no longer "live") top of stack as
holding their result.  This meant the caller could use the value without
also having the responsibility of having to pop it to maintain stack balance.

As the many warnings about "EXTREMELY UNSTABLE" suggest, this is
not a very good idea.  You wind up with the caller needing to be aware
of the stack anyway, only in a much more scary way.  It also means the
stack can't be aggressive about catching stray accesses to invalid stack
positions because "the code might have meant to do that".

This first step gets rid of the TOS1 protocol in function apply routines.