Bring back RE_NO_VALUE omitted in 0948f4f
This line was lost I converted a Get_Var call to a GET_VAR_INTO based on
the mistaken belief that GET_VAR_INTO would trap unsets and give an error.
Whether it should or not is up for debate--it probably should.  At the moment it
does not, and it didn't in the Do_Next replacement I have written.  So just
putting it back for now in this Frankenstein Do_Next (being fabricated for the
sake of 'git bisect' or something).