Make console process "--do" with RL_Do_String
In the Ren/C separation of concerns, things like handling --do on the
command-line are shifted to where the core doesn't know about the
string "--do" or what a "do_arg" is.  Rather, it offers generic evaluation
services.  So I had moved handling of --do out into the host_main.c
early on.

Coming back to it to commit it, the Atronix code has added more
"non-core" stuff into the init script with encapping.  That's another thing
that should fall to each client (and should probably be a little cross
platform library in its own right, generalized to purposes that would
not necessarily have to do with Rebol, if such a library does not
already exist).  One step at a time.

The only *intended* net effect is that since the code is run via the text
of the RL_Do_String, the 'intern' needed in the Rebol version of the
execution never happens.  So the block that b-init.c makes is unbound.
Considering that its likely fate is to not exist at all, that's okay.

Also flattens RL_Do_Binary to a call to RL_Do_String, and some
const-correctness with that.